Rearrange github action
Unit tests require the stopwords dataset, which is handled by dvc.
"Dvc pull" must therefore be run in the same container as pytest.
To remove redundancy "dvc pull" is run before the docker image is built,
so that the data is copied and persisted directly in the container and
can be reused between "docker run .. app" commands.